--- app/views/foreman_ansible/ansible_roles/_ansible_row_buttons.html.erb
@@ -1,15 +1,0 @@
-<td>
-    <%=
-      if hostgroup.all_ansible_roles.empty?
-        'N/A'
-      else
-        button = link_to(
-          icon_text('play', ' ' + _('Ansible roles'), :kind => 'fa'),
-          play_roles_hostgroup_path(:id => hostgroup.id),
-          :id => :ansible_roles_button,
-          :class => 'btn btn-default',
-          :'data-no-turbolink' => true
-        )
-      end
-    %>
-</td>
