--- app/views/foreman_ansible/ansible_roles/_ansible_row_title.html.erb
@@ -1,1 +1,0 @@
-<th><%= _('Ansible Roles') %></th>
